// qcu-app/types/api/products/product.ts
export type ProductsType = {
  product_id: number;
  product_name: string;
  display_name: string;
  display_price: string;
  product_stocks: number;
  product_description: string;
  display_image: string;
  product_sold: number;
  status: string;
  isFeatured: number;
  isDeleted: number;
  date_created: string;
};